// Fitz.Web/src/lib/api/generated/src/lib/api/generated/api-client.ts
export class AdminBuyFitzTicketsRequest implements IAdminBuyFitzTicketsRequest {
    tickets!: number;

    constructor(data?: IAdminBuyFitzTicketsRequest) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.tickets = _data["tickets"];
        }
    }

    static fromJS(data: any): AdminBuyFitzTicketsRequest {
        data = typeof data === 'object' ? data : {};
        let result = new AdminBuyFitzTicketsRequest();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["tickets"] = this.tickets;
        return data;
    }
}

export interface IAdminBuyFitzTicketsRequest {
    tickets: number;
}
// ...
export class AdminExtendLotteryEndDateRequest implements IAdminExtendLotteryEndDateRequest {
    endDate!: Date;

    constructor(data?: IAdminExtendLotteryEndDateRequest) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.endDate = _data["endDate"] ? new Date(_data["endDate"].toString()) : undefined as any;
        }
    }

    static fromJS(data: any): AdminExtendLotteryEndDateRequest {
        data = typeof data === 'object' ? data : {};
        let result = new AdminExtendLotteryEndDateRequest();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["endDate"] = this.endDate ? this.endDate.toISOString() : undefined as any;
        return data;
    }
}

export interface IAdminExtendLotteryEndDateRequest {
    endDate: Date;
}
// ...
export class AdminModifyLotteryPoolRequest implements IAdminModifyLotteryPoolRequest {
    pool!: number;

    constructor(data?: IAdminModifyLotteryPoolRequest) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.pool = _data["pool"];
        }
    }

    static fromJS(data: any): AdminModifyLotteryPoolRequest {
        data = typeof data === 'object' ? data : {};
        let result = new AdminModifyLotteryPoolRequest();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["pool"] = this.pool;
        return data;
    }
}

export interface IAdminModifyLotteryPoolRequest {
    pool: number;
}
// ...
export class BulkUpdateFavorabilityRequest implements IBulkUpdateFavorabilityRequest {
    userIds?: number[] | undefined;
    favorability?: number;

    constructor(data?: IBulkUpdateFavorabilityRequest) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            if (Array.isArray(_data["userIds"])) {
                this.userIds = [] as any;
                for (let item of _data["userIds"])
                    this.userIds!.push(item);
            }
            this.favorability = _data["favorability"];
        }
    }

    static fromJS(data: any): BulkUpdateFavorabilityRequest {
        data = typeof data === 'object' ? data : {};
        let result = new BulkUpdateFavorabilityRequest();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        if (Array.isArray(this.userIds)) {
            data["userIds"] = [];
            for (let item of this.userIds)
                data["userIds"].push(item);
        }
        data["favorability"] = this.favorability;
        return data;
    }
}

export interface IBulkUpdateFavorabilityRequest {
    userIds?: number[] | undefined;
    favorability?: number;
}
// ...
export class EvaluatePollRequestDto implements IEvaluatePollRequestDto {
    status!: PollStatusEnum;

    constructor(data?: IEvaluatePollRequestDto) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.status = _data["status"];
        }
    }

    static fromJS(data: any): EvaluatePollRequestDto {
        data = typeof data === 'object' ? data : {};
        let result = new EvaluatePollRequestDto();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["status"] = this.status;
        return data;
    }
}
// ...
export enum PollStatusEnum {
    Pending = "Pending",
    Approved = "Approved",
    Declined = "Declined",
}

export enum PollTypeEnum {
    Number = "Number",
    YesOrNo = "YesOrNo",
    Color = "Color",
    ThisOrThat = "ThisOrThat",
    HotTake = "HotTake",
}

export enum RenameStatusEnum {
    Unknown = "Unknown",
    Pending = "Pending",
    Active = "Active",
    Expired = "Expired",
    BoughtOut = "BoughtOut",
    Permanent = "Permanent",
}

export class SetLotterySubscribeRequest implements ISetLotterySubscribeRequest {
    userId!: number;
    subscribe!: boolean;

    constructor(data?: ISetLotterySubscribeRequest) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.userId = _data["userId"];
            this.subscribe = _data["subscribe"];
        }
    }

    static fromJS(data: any): SetLotterySubscribeRequest {
        data = typeof data === 'object' ? data : {};
        let result = new SetLotterySubscribeRequest();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["userId"] = this.userId;
        data["subscribe"] = this.subscribe;
        return data;
    }
}

export interface ISetLotterySubscribeRequest {
    userId: number;
    subscribe: boolean;
}

export class SetSafeBalanceRequest implements ISetSafeBalanceRequest {
    userId!: number;
    safeBalance!: number;

    constructor(data?: ISetSafeBalanceRequest) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.userId = _data["userId"];
            this.safeBalance = _data["safeBalance"];
        }
    }

    static fromJS(data: any): SetSafeBalanceRequest {
        data = typeof data === 'object' ? data : {};
        let result = new SetSafeBalanceRequest();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["userId"] = this.userId;
        data["safeBalance"] = this.safeBalance;
        return data;
    }
}

export interface ISetSafeBalanceRequest {
    userId: number;
    safeBalance: number;
}

export class SetTicketAmountRequest implements ISetTicketAmountRequest {
    userId!: number;
    amount!: number;

    constructor(data?: ISetTicketAmountRequest) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.userId = _data["userId"];
            this.amount = _data["amount"];
        }
    }

    static fromJS(data: any): SetTicketAmountRequest {
        data = typeof data === 'object' ? data : {};
        let result = new SetTicketAmountRequest();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["userId"] = this.userId;
        data["amount"] = this.amount;
        return data;
    }
}

export interface ISetTicketAmountRequest {
    userId: number;
    amount: number;
}
// ...
export class UpdateFavorabilityRequest implements IUpdateFavorabilityRequest {
    favorability?: number;

    constructor(data?: IUpdateFavorabilityRequest) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.favorability = _data["favorability"];
        }
    }

    static fromJS(data: any): UpdateFavorabilityRequest {
        data = typeof data === 'object' ? data : {};
        let result = new UpdateFavorabilityRequest();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["favorability"] = this.favorability;
        return data;
    }
}

export interface IUpdateFavorabilityRequest {
    favorability?: number;
}

export class UpdateFavorabilitySettingsRequest implements IUpdateFavorabilitySettingsRequest {
    favorabilityBeerRatioThreshold?: number | undefined;
    favorabilityLowThreshold?: number | undefined;
    favorabilityBaseDropPercent?: number | undefined;
    favorabilityDropMultiplier?: number | undefined;

    constructor(data?: IUpdateFavorabilitySettingsRequest) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.favorabilityBeerRatioThreshold = _data["favorabilityBeerRatioThreshold"];
            this.favorabilityLowThreshold = _data["favorabilityLowThreshold"];
            this.favorabilityBaseDropPercent = _data["favorabilityBaseDropPercent"];
            this.favorabilityDropMultiplier = _data["favorabilityDropMultiplier"];
        }
    }

    static fromJS(data: any): UpdateFavorabilitySettingsRequest {
        data = typeof data === 'object' ? data : {};
        let result = new UpdateFavorabilitySettingsRequest();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["favorabilityBeerRatioThreshold"] = this.favorabilityBeerRatioThreshold;
        data["favorabilityLowThreshold"] = this.favorabilityLowThreshold;
        data["favorabilityBaseDropPercent"] = this.favorabilityBaseDropPercent;
        data["favorabilityDropMultiplier"] = this.favorabilityDropMultiplier;
        return data;
    }
}

export interface IUpdateFavorabilitySettingsRequest {
    favorabilityBeerRatioThreshold?: number | undefined;
    favorabilityLowThreshold?: number | undefined;
    favorabilityBaseDropPercent?: number | undefined;
    favorabilityDropMultiplier?: number | undefined;
}

export class UpdateRenameStatusRequestDto implements IUpdateRenameStatusRequestDto {
    status!: RenameStatusEnum;

    constructor(data?: IUpdateRenameStatusRequestDto) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.status = _data["status"];
        }
    }

    static fromJS(data: any): UpdateRenameStatusRequestDto {
        data = typeof data === 'object' ? data : {};
        let result = new UpdateRenameStatusRequestDto();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["status"] = this.status;
        return data;
    }
}